requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.5.2
python-multipart==0.0.9
pydantic==2.6.4
